Make patch remapping more useable by alloweing to to be resumed
This allows cases where our changes to Spigot's sources that we have to
do to fix cases of bad mappings can be addressed per-patch during the
patch remap process, and continued from.

If an original patch fails to apply, the user can go into the patch
remap directory, fix the patch, run

    git am --continue

and then go back to the Paper directory and restart the process with:

    ./gradlew remapPatches --skip-patches=<number of patch that failed>

The remap patches process will then finish remapping and applying the
patch that failed, and continue with the other patches from there.

Since patch remapping can take so long, this QoL improvement is
essential for allowing Paper forks to be able to remap their own patches
when the time comes.