Handle reobf issues caused by re-compilation and access changes
This commit includes a large comment which describes the problem it
addresses, so I won't re-iterate all of that here. It is important to
note that simply decompiling and recompiling the code can cause these
issues as well, since we tend to lose some of these fine-grain details
during decompilation.