Pass bundler instead of server jar to patcher
Previously the upstream data would point the vanilla jar to the
server.jar the bundler generates. The server jar however does not
contain the metadata required, such as as 'META-INF/versions.list' and
instead only contains the server classes.

This commit changes that behaviour to now pass the bundler jar instead
of the server jar to the upstream data, allowing the patcher to properly
create bundlers.