use already compiled classes as first entry in classpath
Due to the way Paper works, it overwrites many classes
that come from libraries. Since Paper needs the
classes, which it does not overwrite, it still has those
libraries as dependencies. The way, that gradle
constructs the classpath breaks Paper's way of doing
things when it comes to incremental compilation. Thus, we
need to ensure that the compiler always finds our modified
classes before it has a chance to find them in the
dependencies.