Move userdev setup execution to task execution time
Uses a hack where IntelliJ will attach sources inside the same jar as the classes.

This is a working (and from what I can tell configuration cache-compatible) implementation, but is mostly the same pipeline, moved to task execution time. More work is still needed to clean this up and future-proof certain areas.