Rewrite userdev setup pipeline
Uses a graph for ordering steps instead of manual ordering, improved input/output hashing and up-to-date checking.

Except for the final outputs which are keyed to the dev bundle, steps are independent of the dev bundle and can be reused between bundles that have the same inputs for the step. For example when using a Paper and Folia 1.21.4 dev bundle in two projects, decompile will only run once as only the final apply dev bundle patches step differs.

Final outputs are copied into a project local location to avoid projects sharing a bundle from having a task with a shared output location.