Implement Patch Roulette and Git history reconstruction for Minecraft updates (#284)
* Add mechanism for bringing in Git history from an old Minecraft version when updating Minecraft

* Add applyOrMove task for file patches

* Implement patch roulette tasks

* Allow cancelling by path

* Update for patch roulette endpoint changes

* Only revert rejected patch if target file exists

Patches might target files that no longer exist in newer versions.
In such a case, no target file will exist in the mojang source
directory.

* Some logging and todo changes

This might be completely wrong. Awaiting jmp.

* Do not write diff patch rejects during update

We cannot just remove rejectsDir here as the git patches need to be
moved to the reject dir.

* Configurable patch selection size

* Option to reapply patches incase of fail

* Apply each roulette patch individually

In case multiple patches are applied, it is important to apply as many
patches as possible to allow manual resolution of failed applies (e.g.
missing patch file, source file missing, etc).

The alternative is a complete re-apply of all selected patches, which is
unwanted.

* Move reapply opt out

* Add Patch Filters

* fix tests

* now actually fix them

* Move down config and functions

* Fix reapplying patches

* Restructure apply check again

* Correctly update config on single patch cancel

* Output amount of patches found

* move patch roulette api to /api

* initial windows support, use start-patches endpoint instead of loop

* Rename props and fix file separators properly

* fix circular ref

* move PatchSelectionStrategy factory method

---------

Co-authored-by: Bjarne Koll <git@lynxplay.dev>
Co-authored-by: Owen1212055 <23108066+Owen1212055@users.noreply.github.com>
Co-authored-by: Tamion <70228790+notTamion@users.noreply.github.com>
Co-authored-by: MiniDigger | Martin <admin@minidigger.dev>