Allow selecting exactly n patches during roulette (#310)
* Allow selecting exactly n patches during roulette

The package local implementation of the selection strategy works well
for large updates as single packages offer a great "unit" of patches to
work on.

When updating paper to specific snapshots on a weekly basis, generally
the amount of non-applicable patches is a lot lower. Being restricted to
a package in this case is bad dx, as most time is spent selecting more
patches and finishing them, when 10 patches would be a much better
"unit" of patches than a package with a single failed patch.

* Revert local testing diff

* Correctly pass down smaller selection

* Comments