Do not prune tags when rebuilding patches (#341)
A semi-common git config to set globally is fetch.pruneTags which allows
developers to prune tags no longer present on the remote.

This config however breaks patch rebuilding as the file and base tags
are pruned during the pruning fetch of paperweight prior to the
format-patch command.

To prevent this, this commit now explicitly disables patch pruning while
preparing the rebuild.